Extract TopAppBar into its own file
